requirements.txt
aiogram==2.23.1
aiogram-broadcaster==0.0.7
Django==4.1.13
python-decouple==3.8
django-model-utils==4.3.1
django_admin_interface==0.25.0
phonenumbers==8.13.24
requests==2.31.0
openai==1.6.1
apscheduler==3.10.1
